+++ CHANGELOG.md
@@ -10,6 +10,8 @@
         - Fix text layout issues in /reports/…/summary dashboard charts.
     - Admin improvements:
         - Inspectors can set non_public status of reports. #1992
+    - Development improvements:
+        - Add HTML email previewer.
 
 * v2.3.1 (12th February 2018)
     - Front end improvements: